refactor: add entry point in utils
